+++ library/src/main/java/org/dbtools/android/room/CloseableDatabaseWrapperRepository.kt
@@ -17,8 +17,35 @@ import java.io.File
      * class BookDatabaseRepository
      * @Inject constructor(application: Application): CloseableDatabaseWrapperRepository<ShopDatabase>(application) {
      *
-     *   fun getAuthorDao() = getDatabase().authorDao()
-     *   fun getContentDao() = getDatabase().contentDao()
+     *   override fun createDatabase(application: Application, filename: String): BookDatabase {
+     *     return Room.databaseBuilder(application, BookDatabase::class.java, filename)
+     *       .fallbackToDestructiveMigration()
+     *       .build()
+     *   }
+     * }
+ * ```
+ * // Initialize (one time)
+ * bookDatabaseRepository.registerDatabase(application, "book-a", "/sdcard/Downloads/booka")
+ *
+ * ...
+ *
+ * // Regular use
+ * val bookDatabase = bookDatabaseRepository.getDatabase("book-a")
+ * bookDatabase.getAuthorDao().findNameById(123)
+
+ *
+ * ==== OR (with auto registration)
+ *
+ * ```
+     * @Singleton
+     * class BookDatabaseRepository
+     * @Inject constructor(application: Application): CloseableDatabaseWrapperRepository<ShopDatabase>(application) {
+     *
+     *  override fun autoRegisterDatabase(key: String): Boolean {
+     *      val databaseFile = fileUtil.getBookDatabase(key)
+     *      registerDatabase(key, databaseFile.absolutePath)
+     *      return true
+     *  }
      *
      *   override fun createDatabase(application: Application, filename: String): BookDatabase {
      *     return Room.databaseBuilder(application, BookDatabase::class.java, filename)
@@ -28,43 +55,74 @@ import java.io.File
      * }
  * ```
  *
- * // Initialize (one time)
- * bookDatabaseRepository.registerDatabase(application, "book-a", "/sdcard/Downloads/booka")
- *
- * ...
- *
- * // Regular use
+ * // Regular use (getDatabase(...) will auto register "book-a"
  * val bookDatabase = bookDatabaseRepository.getDatabase("book-a")
  * bookDatabase.getAuthorDao().findNameById(123)
+ *
  */
 abstract class CloseableDatabaseWrapperRepository<out T: RoomDatabase>(protected val application: Application) {
     private var databaseList = HashMap<String, T>()
 
     /**
      * Get database from internal repository.
-     *
-     * @param key of the database
-     * @return RoomDatabase or IllegalStateException if the database does not exist
-     */
-    @Synchronized
-    open fun getDatabase(key: String) = databaseList[key] ?: throw IllegalStateException("Database for key [$key] does not exist.  Be sure to registerDatabase(key, filename) before calling getDatabase(key)")
-
-    /**
-     * Get database from internal repository.
+     * NOTE: override autoRegisterDatabase(...) function to allow the getDatabase(...) function to auto register a database
      *
      * @param key of the database
      * @return RoomDatabase or null if the database does not exist
      */
-    @Synchronized
-    open fun getDatabaseOrNull(key: String): T? = databaseList[key]
+    open fun getDatabase(key: String): T? {
+        if (key.isBlank()) {
+            Timber.e("key for the database is unspecified")
+            return null
+        }
+
+        // if database is not registered, then try to auto-register
+        if (!isDatabaseRegistered(key, true)) {
+            Timber.d("database is not registered for key [$key]")
+            return null
+        }
+
+        return databaseList[key]
+    }
+
+    /**
+     * Override this function to allow the getDatabase(...) function to auto register a database.  The override function needs to call registerDatabase(...)
+     *
+     * @param key of the database
+     * @return true if the database was registered
+     */
+    open fun autoRegisterDatabase(key: String) = false
 
     /**
      * Identify if the database has been added to the internal repository
+     * NOTE: override autoRegisterDatabase(...) function to allow the isDatabaseRegistered(...) function to auto register a database
      *
      * @param key of the database
-     * @return true if the database has already been added
+     * @return true if the database is added to the repository
      */
-    open fun isDatabaseRegistered(key: String) = databaseList.containsKey(key)
+    open fun isDatabaseRegistered(key: String, autoRegister: Boolean = true): Boolean {
+        val registered = databaseList.containsKey(key)
+
+        return when {
+            registered -> true
+            !registered && autoRegister -> internalAutoRegister(key)
+            else -> false
+        }
+    }
+
+    /**
+     * This function will call the override function autoRegisterDatabase(...)
+     * @return true if the override function autoRegisterDatabase(...) successfully registered the database (this function will verify if it actually got added AFTER)
+     */
+    @Synchronized
+    private fun internalAutoRegister(key: String): Boolean {
+        if (autoRegisterDatabase(key)) {
+            // make sure it actually got added
+            return databaseList.containsKey(key)
+        }
+
+        return false
+    }
 
     /**
      * Add database to internal repository
@@ -75,7 +133,7 @@ abstract class CloseableDatabaseWrapperRepository<out T: RoomDatabase>(protected
      */
     @Synchronized
     open fun registerDatabase(key: String, filename: String): Boolean {
-        if (isDatabaseRegistered(key)) {
+        if (isDatabaseRegistered(key, false)) {
             Timber.e("Database already registered for key [$key]")
             return false
         }